Revert "#018 Purchase UI (#5)"
This reverts commit 96eb7770485191e40c62f4e8bd1e515344b7ffe1.
