+++ Assets/Code/Components/Core.cs
