+++ Assets/Code/Components/HUDcore.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 3083a22c8a76c0447abcc30522f107e6
+guid: 6554c34e40f41c44db1cd3edebabb135
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
