+++ Assets/Code/Components/Logger.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 0a8f1a0fb52d95b4c85c6f4e8452c9ce
+guid: 155ab223ca51c1b40957309d7bfb57e2
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
