+++ Assets/Code/Components/ProducerVisual.cs
