+++ Assets/Code/Components/TMP_Observe.cs
@@ -14,10 +14,12 @@ public class TMP_Observe : MonoBehaviour
     protected object[] datas;
 
     // inspector fields
-    [SerializeField] [Tooltip("Gameobject.Component.Field")]
-    public string[] URI;
-    [SerializeField] [Tooltip("Score: {0:C2}")]
-    public string format;
+    [Tooltip("Gameobject.Component.Field")]
+    [SerializeField]
+    string[] URI;
+    [Tooltip("Gold: {0:C2}")]
+    [SerializeField]
+    string format;
 
     // Start is called before the first frame update
     void Start()
@@ -33,7 +35,10 @@ public class TMP_Observe : MonoBehaviour
         {
             datas[i] = GetData(URI[i]);
         }
+        Debug.Log("format = " + format);
+        Debug.Log("datas = " + datas.Select(x => x.ToString()).ToArray());
         string text = string.Format(format, datas.Select(x => x.ToString()).ToArray());
+        Debug.Log("text = " + text);
         TMPptr.SetText(text);
     }
