+++ Assets/Code/Components/UpgradeManagerBase.cs
