+++ Assets/Code/Components/UpgradeManagerBase.cs.meta
