--- Assets/Code/GameBackend/ClickerProducer.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 98e7365bfbc789741a03f128b530608b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
