--- Assets/Code/Helpers.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d5d917be8b36c2a43b5a2c207f39a3d0
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
