--- Assets/Code/Helpers/Reflection.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e2974d82ee588304594f248cd9c7165f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
