--- Assets/Code/Helpers/UI.cs
@@ -1,264 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using UnityEngine;
-using UnityEngine.UI;
-using TMPro;
-
-namespace Assets.Code.Helpers
-{
-    public enum AnchorPresets
-    {
-        TopLeft,
-        TopCenter,
-        TopRight,
-
-        MiddleLeft,
-        MiddleCenter,
-        MiddleRight,
-
-        BottomLeft,
-        BottonCenter,
-        BottomRight,
-        BottomStretch,
-
-        VertStretchLeft,
-        VertStretchRight,
-        VertStretchCenter,
-
-        HorStretchTop,
-        HorStretchMiddle,
-        HorStretchBottom,
-
-        StretchAll
-    }
-
-    public enum PivotPresets
-    {
-        TopLeft,
-        TopCenter,
-        TopRight,
-
-        MiddleLeft,
-        MiddleCenter,
-        MiddleRight,
-
-        BottomLeft,
-        BottomCenter,
-        BottomRight,
-    }
-
-    public struct ButtonDescriptor
-    {
-        public GameObject gameObject;
-        public Button btnComp;
-        public TextMeshProUGUI textComp;
-    }
-
-    public static class UI
-    {
-        public static void SetAnchor(this RectTransform source, AnchorPresets align)
-        {
-            switch (align)
-            {
-                case (AnchorPresets.TopLeft):
-                    {
-                        source.anchorMin = new Vector2(0, 1);
-                        source.anchorMax = new Vector2(0, 1);
-                        break;
-                    }
-                case (AnchorPresets.TopCenter):
-                    {
-                        source.anchorMin = new Vector2(0.5f, 1);
-                        source.anchorMax = new Vector2(0.5f, 1);
-                        break;
-                    }
-                case (AnchorPresets.TopRight):
-                    {
-                        source.anchorMin = new Vector2(1, 1);
-                        source.anchorMax = new Vector2(1, 1);
-                        break;
-                    }
-
-                case (AnchorPresets.MiddleLeft):
-                    {
-                        source.anchorMin = new Vector2(0, 0.5f);
-                        source.anchorMax = new Vector2(0, 0.5f);
-                        break;
-                    }
-                case (AnchorPresets.MiddleCenter):
-                    {
-                        source.anchorMin = new Vector2(0.5f, 0.5f);
-                        source.anchorMax = new Vector2(0.5f, 0.5f);
-                        break;
-                    }
-                case (AnchorPresets.MiddleRight):
-                    {
-                        source.anchorMin = new Vector2(1, 0.5f);
-                        source.anchorMax = new Vector2(1, 0.5f);
-                        break;
-                    }
-
-                case (AnchorPresets.BottomLeft):
-                    {
-                        source.anchorMin = new Vector2(0, 0);
-                        source.anchorMax = new Vector2(0, 0);
-                        break;
-                    }
-                case (AnchorPresets.BottonCenter):
-                    {
-                        source.anchorMin = new Vector2(0.5f, 0);
-                        source.anchorMax = new Vector2(0.5f, 0);
-                        break;
-                    }
-                case (AnchorPresets.BottomRight):
-                    {
-                        source.anchorMin = new Vector2(1, 0);
-                        source.anchorMax = new Vector2(1, 0);
-                        break;
-                    }
-
-                case (AnchorPresets.HorStretchTop):
-                    {
-                        source.anchorMin = new Vector2(0, 1);
-                        source.anchorMax = new Vector2(1, 1);
-                        break;
-                    }
-                case (AnchorPresets.HorStretchMiddle):
-                    {
-                        source.anchorMin = new Vector2(0, 0.5f);
-                        source.anchorMax = new Vector2(1, 0.5f);
-                        break;
-                    }
-                case (AnchorPresets.HorStretchBottom):
-                    {
-                        source.anchorMin = new Vector2(0, 0);
-                        source.anchorMax = new Vector2(1, 0);
-                        break;
-                    }
-
-                case (AnchorPresets.VertStretchLeft):
-                    {
-                        source.anchorMin = new Vector2(0, 0);
-                        source.anchorMax = new Vector2(0, 1);
-                        break;
-                    }
-                case (AnchorPresets.VertStretchCenter):
-                    {
-                        source.anchorMin = new Vector2(0.5f, 0);
-                        source.anchorMax = new Vector2(0.5f, 1);
-                        break;
-                    }
-                case (AnchorPresets.VertStretchRight):
-                    {
-                        source.anchorMin = new Vector2(1, 0);
-                        source.anchorMax = new Vector2(1, 1);
-                        break;
-                    }
-
-                case (AnchorPresets.StretchAll):
-                    {
-                        source.anchorMin = new Vector2(0, 0);
-                        source.anchorMax = new Vector2(1, 1);
-                        break;
-                    }
-            }
-        }
-
-        public static void SetPivot(this RectTransform source, PivotPresets preset)
-        {
-
-            switch (preset)
-            {
-                case (PivotPresets.TopLeft):
-                    {
-                        source.pivot = new Vector2(0, 1);
-                        break;
-                    }
-                case (PivotPresets.TopCenter):
-                    {
-                        source.pivot = new Vector2(0.5f, 1);
-                        break;
-                    }
-                case (PivotPresets.TopRight):
-                    {
-                        source.pivot = new Vector2(1, 1);
-                        break;
-                    }
-
-                case (PivotPresets.MiddleLeft):
-                    {
-                        source.pivot = new Vector2(0, 0.5f);
-                        break;
-                    }
-                case (PivotPresets.MiddleCenter):
-                    {
-                        source.pivot = new Vector2(0.5f, 0.5f);
-                        break;
-                    }
-                case (PivotPresets.MiddleRight):
-                    {
-                        source.pivot = new Vector2(1, 0.5f);
-                        break;
-                    }
-
-                case (PivotPresets.BottomLeft):
-                    {
-                        source.pivot = new Vector2(0, 0);
-                        break;
-                    }
-                case (PivotPresets.BottomCenter):
-                    {
-                        source.pivot = new Vector2(0.5f, 0);
-                        break;
-                    }
-                case (PivotPresets.BottomRight):
-                    {
-                        source.pivot = new Vector2(1, 0);
-                        break;
-                    }
-            }
-        }
-
-        // arranges a UI transform inside its parent transform
-        // anchor and pivot are set like in inspector
-        // offset is a multiplier of the parent width and height which determins the anchor position
-        // widthMult and heightMult determine width and height as multiplier of parent width and height respectively
-        public static void Arrange(this RectTransform source, RectTransform parent, AnchorPresets anchor, PivotPresets pivot, Vector2 anchorOffsetMult, Vector2 sizeMult)
-        {
-            source.SetParent(parent, false);
-            Vector2 parentSize = new Vector2(parent.rect.width, parent.rect.height);
-            source.anchoredPosition = anchorOffsetMult * parentSize;
-            source.sizeDelta = sizeMult * parentSize;
-            source.SetAnchor(anchor);
-            source.SetPivot(pivot);
-        }
-
-        public static ButtonDescriptor GenerateButton(RectTransform parent, string name = "Button", string text = "")
-        {
-            return GenerateButton(parent, name, AnchorPresets.TopLeft, PivotPresets.TopLeft, Vector2.zero, Vector2.one, text);
-        }
-
-        public static ButtonDescriptor GenerateButton(RectTransform parent, string name, 
-            AnchorPresets anchor, PivotPresets pivot, Vector2 offsetMult, Vector2 sizeMult, string text = "")
-        {
-            // return type
-            ButtonDescriptor ret;
-
-            // make button
-            ret.gameObject = new GameObject(name, typeof(RectTransform));
-            ((RectTransform)ret.gameObject.transform).Arrange(parent, anchor, pivot, offsetMult, sizeMult);
-            ret.btnComp = ret.gameObject.AddComponent<Button>();
-
-            // subobject for text
-            GameObject textGameObject = new GameObject("ButtonText", typeof(RectTransform));
-            ((RectTransform)textGameObject.transform).Arrange((RectTransform)ret.gameObject.transform, AnchorPresets.MiddleCenter, PivotPresets.MiddleCenter, Vector2.zero, Vector2.one);
-            ret.textComp = textGameObject.AddComponent<TextMeshProUGUI>();
-            ret.textComp.SetText(text);
-
-            return ret;
-        }
-    }
-}
