--- Assets/Code/Helpers/UI.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2747a806df9091f47b0539a050c3c73a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
