+++ Assets/GUI.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 6654bbc43e60e6b469b0f38e06764ba7
+guid: a629cda8b31792240a145ba6a2fe4516
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
