--- Assets/SerializableDictionary.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 35b2c65d5ed3e584abb161351ac41697
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
