--- Assets/SerializableDictionary/Editor.meta
@@ -1,5 +1,0 @@
-fileFormatVersion: 2
-guid: 20029ee1c099f4b47afbac48208879b9
-folderAsset: yes
-DefaultImporter:
-  userData:
