--- Assets/SerializableDictionary/Editor/SerializableDictionaryPropertyDrawer.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 91da51d02ab9ebc459d80d5965d40d19
-timeCreated: 1492869349
-licenseType: Store
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
