--- Assets/SerializableDictionary/SerializableDictionary.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: e7be1c9624387604fba4005ccf7dbd5a
-timeCreated: 1492868176
-licenseType: Store
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
