--- Assets/SerializableDictionary/SerializableDictionary.pdf.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 8eb73a56b3a6ee64ab2c30cac8878d8f
-timeCreated: 1492869030
-licenseType: Store
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
